backend/pyproject.toml
[project]
name = "love-talk-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlmodel>=0.0.14",
    "asyncpg>=0.29.0",
    "greenlet>=3.0.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.26.0",
    "pydantic-settings>=2.1.0",
    "email-validator>=2.1.0",
    "alembic>=1.13.0",
    "python-multipart>=0.0.6",
    "psycopg2-binary>=2.9.11",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100
exclude = [
    "packages/db/migrations/versions/*",  # Auto-generated migration files
    "packages/repositories/*",  # Repository files with syntax issues
]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501", "B008"]  # B008: Depends() calls in function defaults are FastAPI pattern

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
exclude = "packages/repositories/.*|packages/db/migrations/versions/.*"

[[tool.mypy.overrides]]
module = "packages.repositories.*"
ignore_errors = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
